bus/hp_dio/98620: convert to anonymous namespace (nw)
